KOGITO-202 avoid open docks by default on kogito editors
