+++ kie-wb-common-stunner/kie-wb-common-stunner-extensions/kie-wb-common-stunner-submarine/kie-wb-common-stunner-submarine-client/src/main/java/org/kie/workbench/common/stunner/submarine/client/docks/DiagramEditorPropertiesDock.java
@@ -22,93 +22,34 @@ import org.gwtbootstrap3.client.ui.constants.IconType;
 import org.jboss.errai.ui.client.local.spi.TranslationService;
 import org.kie.workbench.common.stunner.submarine.client.resources.i18n.SubmarineClientMessages;
 import org.kie.workbench.common.stunner.submarine.client.screens.DiagramEditorPropertiesScreen;
-import org.uberfire.client.workbench.docks.UberfireDock;
-import org.uberfire.client.workbench.docks.UberfireDockPosition;
 import org.uberfire.client.workbench.docks.UberfireDocks;
-import org.uberfire.mvp.impl.DefaultPlaceRequest;
 
 @ApplicationScoped
-public class DiagramEditorPropertiesDock {
-
-    static final double DOCK_SIZE = 400d;
-
-    private UberfireDocks uberfireDocks;
-
-    private TranslationService translationService;
-
-    private UberfireDock uberfireDock;
-
-    private boolean isOpened = false;
-
-    private String owningPerspectiveId;
+public class DiagramEditorPropertiesDock extends BaseDiagramEditorDock {
 
     public DiagramEditorPropertiesDock() {
         // CDI proxy
+        this(null, null);
     }
 
     @Inject
     public DiagramEditorPropertiesDock(final UberfireDocks uberfireDocks,
                                        final TranslationService translationService) {
-        this.uberfireDocks = uberfireDocks;
-        this.translationService = translationService;
+        super(uberfireDocks, translationService);
     }
 
-    public void init(final String owningPerspectiveId) {
-        this.owningPerspectiveId = owningPerspectiveId;
-        this.uberfireDock = makeUberfireDock();
-    }
-
-    public void open() {
-        if (isOpened()) {
-            return;
-        }
-
-        isOpened = true;
-        uberfireDocks.add(getUberfireDock());
-        uberfireDocks.show(position(), owningPerspectiveId());
-        uberfireDocks.open(getUberfireDock());
-    }
-
-    public void close() {
-        if (!isOpened()) {
-            return;
-        }
-
-        isOpened = false;
-        uberfireDocks.close(getUberfireDock());
-        uberfireDocks.remove(getUberfireDock());
-    }
-
-    boolean isOpened() {
-        return isOpened;
-    }
-
-    UberfireDock makeUberfireDock() {
-        final UberfireDock uberfireDock = new UberfireDock(position(), icon(), placeRequest(), owningPerspectiveId());
-        return uberfireDock.withSize(DOCK_SIZE).withLabel(dockLabel());
-    }
-
-    UberfireDockPosition position() {
-        return UberfireDockPosition.EAST;
-    }
-
-    private String icon() {
+    @Override
+    final String icon() {
         return IconType.PENCIL_SQUARE_O.toString();
     }
 
-    private DefaultPlaceRequest placeRequest() {
-        return new DefaultPlaceRequest(DiagramEditorPropertiesScreen.SCREEN_ID);
+    @Override
+    final String getScreenId() {
+        return DiagramEditorPropertiesScreen.SCREEN_ID;
     }
 
-    String owningPerspectiveId() {
-        return owningPerspectiveId;
-    }
-
-    UberfireDock getUberfireDock() {
-        return uberfireDock;
-    }
-
-    private String dockLabel() {
-        return translationService.getTranslation(SubmarineClientMessages.PROPERTIES_DOCK_TITLE);
+    @Override
+    final String getLabelKey() {
+        return SubmarineClientMessages.PROPERTIES_DOCK_TITLE;
     }
 }
